Workout and Exercises Endpoint Changes

+++ src/app/api/workouts/route.ts
@@ -1,29 +1,60 @@
 import connectDB from "@/database/db";
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 
-export async function GET() {
+
+//Search all Workouts
+export async function GET(req: NextRequest) {
   try {
     const db = connectDB();
+    const url = new URL(req.url ? req.url : "invalid");
+    const username = url.searchParams.get("username");
+    const duration = parseInt(url.searchParams.get("duration") || "");
+    const difficulty = url.searchParams.get("difficulty");
+    const tags = url.searchParams.get("tags");
+    const w_name = url.searchParams.get("w_name");
 
-    if (!db) {
-      throw new Error("Failed to connect to the database");
+    let data, error;
+
+    let query = db.from("workout_info").select("*");
+
+    if (username) {
+      query = query.ilike("username", username);
+    }
+    if (!isNaN(duration)) {
+      query = query.eq("duration", duration);
+    }
+    if (difficulty) {
+      query = query.ilike("difficulty", difficulty);
+    }
+    if (tags) {
+      query = query.contains("tags", [tags]);
+    }
+    if (w_name) {
+      query = query.ilike("w_name", w_name);
     }
 
-    const { data: workouts, error } = await db.from("workout_info").select("*");
+    ({ data, error } = await query);
 
     if (error) {
       throw error;
     }
-    // Return the workouts in the response
+
+    if (data.length === 0) {
+      return NextResponse.json({
+        message: "No workouts found matching the criteria",
+        status: 404,
+      });
+    }
+
     return NextResponse.json({
       message: "Workouts retrieved",
       status: 200,
-      workouts,
+      data,
     });
   } catch (error) {
     console.error(error);
     return NextResponse.json({
-      message: `${error}`,
+      message: `Failed to retrieve workouts. ${error}`,
       status: 500,
     });
   }
@@ -38,9 +69,13 @@ export async function POST(req: Request) {
     }
 
     const { username, duration, difficulty, tags, w_name } = await req.json();
+
+    if (!username || !duration || !difficulty || !tags || !w_name) {
+      throw new Error("Missing required parameters");
+    }
+
     console.log(username, duration, difficulty, tags, w_name);
 
-    // Insert the new workout into the Supabase database
     const { data: workout_info, error } = await db
       .from("workout_info")
       .insert([{ username, duration, difficulty, tags, w_name }] as any);
@@ -49,50 +84,87 @@ export async function POST(req: Request) {
       throw error;
     }
 
-    // Return the workout_id in the response
     return NextResponse.json({
       message: "Workout Created",
       status: 201,
       workout_info,
     });
   } catch (error) {
-    console.error(error); // Log the error to the console
+    console.error(error);
     return NextResponse.json({
-      message: `Failed to create workout: ${error}`,
+      message: `Failed to create workout. ${error}`,
       status: 500,
     });
   }
 }
 
+//Delete Workout
 export async function DELETE(req: Request) {
   try {
     const db = connectDB();
+    const url = new URL(req.url ? req.url : "invalid");
+    const w_id = url.searchParams.get("w_id");
 
     if (!db) {
       throw new Error("Failed to connect to the database");
     }
 
-    const { w_id } = await req.json();
-
-    // Delete workout in Supabase database
-    const { data, error } = await db
+    const { data: existingWorkouts, error: queryError } = await db
       .from("workout_info")
-      .delete()
-      .match({ w_id: w_id });
+      .select("*")
+      .eq("w_id", w_id);
 
-    if (error) {
-      throw error;
+    if (existingWorkouts.length == 0) {
+      throw new Error(`Workout with ID ${w_id} does not exist for user`);
     }
 
-    // Return success message in response
+    const { error: deleteError } = await db
+      .from("workout_info")
+      .delete()
+      .match({ w_id });
+
+    if (deleteError) {
+      throw deleteError;
+    }
+
     return NextResponse.json({
       message: "Workout deleted",
       status: 200,
     });
   } catch (error) {
-    console.error(error); // Log the error to the console
+    console.error(error);
     return NextResponse.json({
-      message: `Failed to delete workout. Please try again later`,
+      message: `Failed to delete workout. ${error}`,
+      status: 500,
+    });
+  }
+}
+
+// Update Workout
+export async function PUT(req: NextRequest) {
+  try {
+    const db = connectDB();
+
+    const { w_id, username, duration, difficulty, tags, w_name } =
+      await req.json();
+
+    const { error } = await db
+      .from("workout_info")
+      .update({ username, duration, difficulty, tags, w_name })
+      .match({ w_id });
+
+    if (error) {
+      throw error;
+    }
+
+    return NextResponse.json({
+      message: "Workout updated",
+      status: 200,
+    });
+  } catch (error) {
+    console.error(error);
+    return NextResponse.json({
+      message: `Failed to update workout. ${error}`,
       status: 500,
     });
   }
